requirements.txt
asgiref==3.8.1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-appconf==1.0.6
django-encrypted-model-fields==0.6.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycrypto==2.6.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-keyczar==0.716
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
